+++ module-karel/src/partitions.ts
@@ -1,5 +1,5 @@
 import { Submission } from "classroom-api"
-import { HwConfig } from "./config"
+import { HwConfig } from "./homework"
 
 type S = Submission
 export interface Partitions<T> {
