+++ module-karel/src/runs.ts
@@ -1,7 +1,8 @@
 import fs from 'fs'
 import {Submission} from 'classroom-api'
 import {Partitions} from './partitions'
-import {HwConfig} from './config'
+import {HwConfig} from './homework'
+import path from 'path'
 
 export interface RunOpts {
     trial?: boolean,
@@ -16,7 +17,7 @@ export function log<T>(e: T, message: string) {
     console.log(message)
     return e
 }
-const data_path = `/home/gigi/Desktop/freeuni-digital-technologies/subject-modules/data`
+const data_path = path.resolve(__dirname, `../../../data`)
 const results_path = `${data_path}/output`
 const submissions_path = `${data_path}/submissions`
