+++ module-karel/src/subject.ts
@@ -1,0 +1,3 @@
+// ეს ფაილი წაიკითხავს საგნის კონფიგურაციას
+// შესაძლებელი უნდა იყოს სხვადასხვა config optionები რომ 
+// ერთ ფაილში იყოს გაერთიანებული
